# Conflicts:
#	LunaDll/GlobalFuncs.cpp
#	LunaDll/LuaMain/LuaProxyComponent/LuaProxyGlobalFunctions/LuaProxyGlobalFuncGraphics.cpp
#	LunaDll/LuaMain/LuaProxyFFI.cpp
#	LunaDll/LuaMain/LunaLuaMain.cpp
#	LunaDll/Misc/RuntimeHookComponents/RuntimeHookCharacterId.cpp
#	LunaDll/Misc/RuntimeHookComponents/RuntimeHookHooks.cpp
#	LunaDll/Rendering/GL/GLEngineCmds.h
#	LunaDll/Rendering/GL/GLTextureStore.cpp

+++ LunaDll/LuaMain/LuaProxyComponent/LuaProxyShader.cpp
@@ -68,7 +68,7 @@ luabind::object convertShaderVariableInfo(const GLShaderVariableInfo* info, lua_
     infoTbl["name"] = info->getName();
     infoTbl["rawName"] = info->getRawName();
     infoTbl["arrayCount"] = info->arrayCount();
-    infoTbl["type"] = info->getType();
+    infoTbl["type"] = static_cast<unsigned int>(info->getType());
     infoTbl["varInfoType"] = static_cast<int>(info->getVarType());
     infoTbl["arrayDepth"] = info->getArrayDepth();
     return infoTbl;
